updated swagger.yaml for multiple schemas
