+++ api/controller/delete.js
@@ -4,7 +4,7 @@ const beautify = require('js-beautify').js;
 module.exports = ({ schema, logging, destination, name }) => {
   const action = 'delete';
   const { uppercase, createUpdateCode, sugarGenerated } = require('../utils');
-  if (logging) console.log(`API => REST => DELETE ${name}`);
+  if (logging) console.log(`  API => REST => DELETE ${name}`);
   const controllerSubFolder = `${destination}/controller/${name}`;
   const createFile = `${controllerSubFolder}/delete${uppercase(name)}.js`;
