+++ api/swagger/index.js
@@ -3,4 +3,6 @@ module.exports = {
   safeDefault: require('./safeDefault'),
   schemaDefinition: require('./schemaDefinition'),
   makeProperties: require('./makeProperties'),
+  makeSwaggerModelDefinitions: require('./makeSwaggerModelDefinitions'),
+
 };
